* fixed bug in reset logic * moved ZeroInSyn callback down into Compiler as it's helpful for lots of compiler types
